Update about and remove /now
I may be overthinking this, but the effort needed to keep up the /now
page does not produce the value that I would like. My main value add in
terms of this blog are the writings about technical topics, specifically
developing backend software systems, and about programming languages,
and also about human language. I want to spend as much time doing and
creating as opposed keeping up meta stuff, such as writing about myself
in ways that don't add value to either myself or others. In other words,
it's easy to say what you're trying to do, but it's harder to actually
do it and produce tangible results.

Wow, that's a long commit message. I hope nobody reads this.