Use border instead of shadow around post images and adjust margin
I'm switching to a border because it goes with the theme of "I want my
blog to look as much like a book as possible", and I adjusted the
margins because the larger margin is no longer needed without the
shadow, as the shadow caused the image to appear closer to the text than
it really was. Also, without the extra margin, the images are now like
paragraphs in terms of layout, which flows better.