Adjust line height and underline offset on home page
The underline position "under" causes some problems in paragraphs in
that the underline is too far away from the word, almost closer to the
line below. To fix this, it requires a large line height which makes the
lines too loose, and thus I've made the underlines closer with my own
offset and made the line height not as loose.