Fix CSS changes requiring local Hugo server restart
After much debugging (to be written up in a post soon!) I found that the
cause of having to restart the local Hugo server every time I make a CSS
change, despite Tailwind CSS building correctly and the files before
`fingerprint` seeming to be all correct, was a change in Hugo where CSS
files are now expected to be in the `assets/` directory and not
`static/`. Having to add that line to config.toml should have been a
clue back then that something was wrong!

This change has the added benefit of all other CSS files besides the
file, fingerprinted one not showing up in the Hugo build directory.