Remove email sign up from work journal
Doesn’t really make sense to ask for this before someone has made it to
the bottom (assuming audience is a new reader). If someone is an
existing reader who keeps coming back, if they really want the updates
via email, I suppose they will come across the subscribe link at the
bottom of an article or the list, at some point. Plus, the email
newsletter isn’t exactly about following the journey of the work journal
from day one, so it’s a bit misleading.