Use .RelPermalink instead of .Permalink
For some reason that I couldn’t quite figure out but nevertheless was
happening, sometimes the site would deploy with the localhost:1313
permalinks there instead of the production one, johnjago.com.

The full link isn’t necessary so I changed it to .RelPermalink in a few
places that matter most, and where it makes sense. (Cannot do this for
the canonical tag, so will need to keep an eye on it.)