Reduce space between items in article lists
This also refactors them to be in `<ul>`s instead of `<div>`s as it’s
more semantic, and it corrects a previous error where I had two closing
`<div>`s and no opening ones.

The space reduction is mainly for the home page, so that the gap between
the articles is no larger than between the heading of the section and
the first article. Any more space and it tends to look off.