Remove day from date in blog and technical post lists
It doesn’t serve the reader to know exactly which day a post was
published when browsing the archives. Knowing at least the month is
enough, so this allows me to reduce some clutter in the list.

The `datetime` attribute retains the full date, as does the post itself
when opened.