Make link styles consistent everywhere and not bold in prose
The color was slightly different in certain sizes of prose. It was also
bold, which when reading, came across as emphasis. Now, the color is
consistent everywhere, and the links are no longer of medium weight in
prose—they are the same normal weight as the surrounding text, so it
does not read as emphasis.

To make the links at least somewhat visible without the bolding, the
underline for links is darker, giving a 4.64 contrast against the
background.