Try out using a section to contain a series
I moved the only “series” of posts that I had into a Hugo Section, using
that to mark those posts as being part of a series, which I then use to
display previous and next post links at the bottom of each post in the
series, as well as a link to the page explaining the series next to the
date.

This has the added benefit of cleaning up the content/ directory a
little.