Improve font loading
For some reason, I had the query parameters for weights and italics not
entered in the same way that Google has in their snippet in Google
Fonts, which was causing it to make 3 identical network requests for
Source Sans 3.

By using their official snippet, all weights and italics are loaded, and
only 1 network request is made, reducing data transfer size by 66%.

Also, I found that you can request a font for just certain characters,
so I updated the query parameters for Pacifico to only request the
letters I use Pacifico for (just for the blog title).

In all, this reduced the transfer size for fonts from 110.65 kB to 32 kB.